requirements.txt
# Server
loguru
fastapi
uvicorn
# Encode
vdf
ujson
# Web
jinja2
aiohttp
pywebview
websockets
# Any things
Pillow
pystray